erictsu/requirements/common.txt
aiohttp==0.17.2
aiomysql==0.0.4
alabaster==0.7.6
alembic==0.8.0
argcomplete==0.9.0
astropy==1.0.3
Babel==1.3
bcolz==0.9.0
beautifulsoup4==4.4.0
binstar==0.11.0
bitarray==0.8.1
blaze==0.8.2
bleach==1.4.1
blinker==1.4
blz==0.6.2
bokeh==0.9.0
boto==2.38.0
Bottleneck==1.0.0
certifi==2015.4.28
cffi==1.1.2
chardet==2.3.0
click==5.1
clyent==0.3.4
colorama==0.3.3
conda==3.16.0
conda-build==1.14.1
conda-env==2.4.2
configobj==5.0.6
coverage==3.7.1
cryptography==0.9.1
cssselect==0.9.1
Cython==0.22.1
cytoolz==0.7.3
datashape==0.4.6
decorator==4.0.2
defusedxml==0.4.1
django==1.8.3
docutils==0.12
dominate==2.1.12
fastcache==1.0.2
feedparser==5.2.1
Flask==0.10.1
Flask-Babel==0.9
Flask-Bootstrap==3.3.5.7.dev1
Flask-HTTPAuth==2.6.0
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.5.0
Flask-Moment==0.5.1
Flask-OpenID==1.2.4
Flask-PageDown==0.2.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
Flasky==0.1.0
flipflop==1.0
greenlet==0.4.7
guess-language==0.2
gunicorn==19.1.1
h5py==2.5.0
html5lib==0.999999
idna==2.0
ipykernel==4.0.3
ipython==4.0.0
ipython-genutils==0.1.0
ipywidgets==4.0.2
itsdangerous==0.24
jdcal==1.0
jedi==0.9.0
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.0.0
jupyter-console==4.0.1
jupyter-core==4.0.4
llvmlite==0.5.0
lxml==3.4.4
Mako==1.0.1
Markdown==2.6.2
MarkupSafe==0.23
matplotlib==1.4.3
menuinst==1.0.4
mistune==0.7.1
mock==1.3.0
multipledispatch==0.4.8
mysql-connector-python==2.0.3
nbconvert==4.0.0
nbformat==4.0.0
networkx==1.10
nltk==3.0.3
nose==1.3.7
notebook==4.0.4
numba==0.19.1
numexpr==2.4.3
numpy==1.9.2
odo==0.3.3
openpyxl==2.2.5
pandas==0.16.2
path.py==0.0.0
patsy==0.4.0
pbr==0.10.0
pep8==1.6.2
pickleshare==0.5
Pillow==2.8.2
ply==3.6
psutil==3.1.1
py==1.4.30
pyasn1==0.1.8
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5.1
pyflakes==0.9.2
Pygments==2.0.2
PyMySQL==0.6.6
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyquery==1.2.9
pyreadline==2.0
pyspider==0.3.5
pytest==2.7.2
python-dateutil==2.4.2
python-editor==0.3
python-openid==2.2.5
python3-openid==3.0.6
pytz==2014.7
pywin32==219
PyYAML==3.11
pyzmq==14.7.0
qtconsole==4.0.1
requests==2.7.0
rope-py3k==0.9.4.post1
runipy==0.1.3
sae-python-dev==1.3.3
scikit-image==0.11.3
scikit-learn==0.16.1
scipy==0.15.1
simplegeneric==0.8.1
six==1.8.0
snowballstemmer==1.2.0
sockjs-tornado==1.0.1
speaklater==1.3
sphinx==1.3.1
sphinx-rtd-theme==0.1.7
spyder==2.3.5.2
SQLAlchemy==0.9.7
sqlalchemy-migrate==0.9.2
sqlparse==0.1.11
statsmodels==0.6.1
sympy==0.7.6
tables==3.2.0
Tempita==0.5.2
toolz==0.7.2
tornado==4.2.1
traitlets==4.0.0
u-msgpack-python==2.1
ujson==1.33
urllib3==1.11
virtualenv==13.0.1
visitor==0.1.2
Werkzeug==0.10.4
wheel==0.24.0
Whoosh==2.6.0
WTForms==2.0.2
xlrd==0.9.4
XlsxWriter==0.7.3
xlwings==0.3.6
xlwt==1.0.0
